fix: TLS or not determination
The `tls` viper option would default to `false`, so if the `--tls`
option was not provided the condition would always default to no-TLS,
i.e. plain text HTTP. We want to be governed by the `tls` flag only if
it was provided, be it `--tls=false` or `--tls=true`. Inherits the
assumption that local loopback listeners are using plain text HTTP. Adds
tests.