refactor: code for conforma compatibility
This commit contains some small refactors for conforma compatibility.
Primary among these are a reversion of the git `Gather()` function and
associated test updates. The `Gather()` function is previously known
working code.

Bits of the `processUrl()` function in the `Git` gather package were
refactored to be more robust.

This commit adds a `containsOCIRegistry()` function to the `OCI`
gatherer package which is used in the oci `Matcher()` function.

Ref: EC-960

Signed-off-by: Rob Nester <rnester@redhat.com>